@@ -1,12 +1,12 @@
 name: cuckoo_workshop
 description: A workshop for Flutter beginners.
 
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -17,7 +17,6 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-
 
 flutter:
   uses-material-design: true
@@ -41,4 +40,3 @@ flutter:
         - asset: assets/fonts/Montserrat-Regular.otf
         - asset: assets/fonts/Montserrat-Bold.ttf
           weight: 700
-
