@@ -19,16 +19,25 @@ class MyApp extends StatelessWidget {
     return MaterialApp(
       title: 'Cuckoo Workshop',
       theme: ThemeData(
-        primarySwatch: MaterialColor(0xFF00F8AD, CustomColors.neonGreenMaterialSwatch),
+        primarySwatch:
+            MaterialColor(0xFF00F8AD, CustomColors.neonGreenMaterialSwatch),
         textTheme: TextTheme(
-          headline1: TextStyle(fontSize: 50.0, fontFamily: 'Pangram-Black', color: Colors.black),
-          headline2: TextStyle(fontSize: 50.0, fontFamily: 'Pangram-Black', color: Colors.white),
-          headline3: TextStyle(fontSize: 30.0, fontFamily: 'Pangram-Black', color: Colors.black),
-          headline4: TextStyle(fontSize: 30.0, fontFamily: 'Pangram-Black', color: Colors.white),
-          bodyText1: TextStyle(fontSize: 14.0, fontFamily: 'Montserrat', color: Colors.white),
-          bodyText2: TextStyle(fontSize: 14.0, fontFamily: 'Montserrat', color: Colors.black),
-          subtitle1: TextStyle(fontSize: 16.0, fontFamily: 'OverpassMono', color: Colors.black),
-          subtitle2: TextStyle(fontSize: 14.0, fontFamily: 'OverpassMono', color: Colors.black),
+          displayLarge: TextStyle(
+              fontSize: 50.0, fontFamily: 'Pangram-Black', color: Colors.black),
+          displayMedium: TextStyle(
+              fontSize: 50.0, fontFamily: 'Pangram-Black', color: Colors.white),
+          displaySmall: TextStyle(
+              fontSize: 30.0, fontFamily: 'Pangram-Black', color: Colors.black),
+          headlineMedium: TextStyle(
+              fontSize: 30.0, fontFamily: 'Pangram-Black', color: Colors.white),
+          bodyLarge: TextStyle(
+              fontSize: 14.0, fontFamily: 'Montserrat', color: Colors.white),
+          bodyMedium: TextStyle(
+              fontSize: 14.0, fontFamily: 'Montserrat', color: Colors.black),
+          titleMedium: TextStyle(
+              fontSize: 16.0, fontFamily: 'OverpassMono', color: Colors.black),
+          titleSmall: TextStyle(
+              fontSize: 14.0, fontFamily: 'OverpassMono', color: Colors.black),
         ),
       ),
       initialRoute: '/',
