@@ -1,7 +1,8 @@
-import 'package:cuckoo_workshop/utilities/custom_colors.dart';
 import 'package:flutter/material.dart';
 
 class AnimateRoute extends StatefulWidget {
+  const AnimateRoute({Key? key}) : super(key: key);
+
   @override
   _AnimateRouteState createState() => _AnimateRouteState();
 }
@@ -14,7 +15,7 @@ class _AnimateRouteState extends State<AnimateRoute> {
     return Scaffold(
       backgroundColor: Colors.black,
       appBar: AppBar(
-        title: Text('Animate', style: Theme.of(context).textTheme.headline3),
+        title: Text('Animate', style: Theme.of(context).textTheme.displaySmall),
       ),
       body: AnimatedOpacity(
         opacity: buttonClicked ? 0.0 : 1.0,
@@ -22,39 +23,39 @@ class _AnimateRouteState extends State<AnimateRoute> {
         child: Column(
           children: [
             Container(
-              width: MediaQuery.of(context).size.width,
+              width: MediaQuery.sizeOf(context).width,
               padding: const EdgeInsets.only(top: 70.0, bottom: 30.0),
               child: Center(
                 child: Image.asset(
                   'assets/pics/cuckoo.png',
-                  width: MediaQuery.of(context).size.width / 2,
+                  width: MediaQuery.sizeOf(context).width / 2,
                   fit: BoxFit.cover,
                 ),
               ),
             ),
             Text(
               "Everything fades away eventually....",
-              style: Theme.of(context).textTheme.bodyText1,
+              style: Theme.of(context).textTheme.bodyLarge,
             ),
             Text(
               "And returns.",
-              style: Theme.of(context).textTheme.bodyText1,
+              style: Theme.of(context).textTheme.bodyLarge,
             ),
             Text(
               "And fades.",
-              style: Theme.of(context).textTheme.bodyText1,
+              style: Theme.of(context).textTheme.bodyLarge,
             ),
             Text(
               "And returns.",
-              style: Theme.of(context).textTheme.bodyText1,
+              style: Theme.of(context).textTheme.bodyLarge,
             ),
             Text(
               "And fades.",
-              style: Theme.of(context).textTheme.bodyText1,
+              style: Theme.of(context).textTheme.bodyLarge,
             ),
             Text(
               "Ah whatever...",
-              style: Theme.of(context).textTheme.bodyText1,
+              style: Theme.of(context).textTheme.bodyLarge,
             ),
           ],
         ),
